etl/loader_postgres.py
"""
PostgreSQL Data Loader
Loads transformed data into PostgreSQL warehouse
"""

import logging
from typing import List, Dict
from sqlalchemy import create_engine, text
from sqlalchemy.orm import sessionmaker
import psycopg2

logger = logging.getLogger(__name__)


class PostgresLoader:
    """Handles data loading into PostgreSQL"""
    
    def __init__(self, connection_string: str):
        self.engine = create_engine(connection_string)
        self.Session = sessionmaker(bind=self.engine)
        
    def load_products(self, products: List[Dict], batch_size: int = 1000):
        """
        Bulk load products into PostgreSQL
        
        Args:
            products: List of product dictionaries
            batch_size: Number of records per batch
        """
        # TODO: Implement bulk insert using COPY or batch inserts
        logger.info(f"Loading {len(products)} products into PostgreSQL")
        pass
    
    def load_price_history(self, price_records: List[Dict], batch_size: int = 1000):
        """
        Bulk load price history records
        
        Args:
            price_records: List of price history dictionaries
            batch_size: Number of records per batch
        """
        # TODO: Implement bulk insert for price history
        pass
    
    def load_reviews(self, review_records: List[Dict], batch_size: int = 1000):
        """
        Bulk load review/sales records
        
        Args:
            review_records: List of review dictionaries
            batch_size: Number of records per batch
        """
        # TODO: Implement bulk insert for reviews
        pass
    
    def incremental_load(self, new_data: Dict[str, List[Dict]]):
        """
        Perform incremental load (upsert) for new/updated records
        
        Args:
            new_data: Dictionary with 'products', 'price_history', 'reviews'
        """
        # TODO: Implement upsert logic
        pass

