requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pyyaml>=6.0

# Database Drivers
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
pymongo>=4.5.0

# ETL & Data Processing
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0

# Scheduling
apscheduler>=3.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
tqdm>=4.66.0
python-dateutil>=2.8.0

